Time: 2 ms (71.36%), Space: 6.2 MB (98.98%) - LeetHub

+++ 1884-egg-drop-with-2-eggs-and-n-floors/1884-egg-drop-with-2-eggs-and-n-floors.cpp
@@ -3,7 +3,7 @@ public:
     int twoEggDrop(int n) {
         
         int i=1;
-        for(i=1; (int)(i*(i+1)/2 )<n  ;i++);
+        for(i=1; (i*(i+1)/2 )<n  ;i++);
         return i;
     }
 };
